+++ src/interfaces.ts
@@ -70,6 +70,7 @@ export interface ServiceType {
   setCharacteristic?: (iid: number, value: number | string | boolean) => Promise<ServiceType>;
   setCharacteristicByType?: (type: string, value: number | string | boolean) => Promise<ServiceType>;
   getCharacteristic?: (type: string) => CharacteristicType;
+  getResource: (body: ResourceRequestType) => Promise<ServiceType>;
   values: any;
   instance: HapInstance;
   uniqueId?: string;
@@ -103,4 +104,11 @@ export interface AccessoryInformationProperties {
   'Name': string;
   'Serial Number': string;
   'Firmware Revision': string;
+}
+
+export interface ResourceRequestType {
+  aid?: number,
+  "resource-type": string,
+  "image-width": number,
+  "image-height": number
 }